Fix build process

+++ bin/configure.js
@@ -29,16 +29,22 @@ console.log(`Loading configuration for project "${pkg.name}"`);
 
 const defaultGitConfigurationUrl = 'https://github.com/wireapp/wire-web-config-default';
 const gitConfigurationUrl = process.env.WIRE_CONFIGURATION_REPOSITORY || defaultGitConfigurationUrl;
-const configDirName = 'config';
+const configDirName = process.env.WIRE_CONFIGURATION_EXTERNAL_DIR || 'config';
 const configDir = resolve(configDirName);
 const src = resolve(configDir, pkg.name, 'content');
 const root = '.';
 const dest = `../../../dist/templates`;
 const ignoreList = ['.DS_Store'];
 
-console.log(`Cleaning config directory "${configDir}"`);
-fs.removeSync(configDir);
-execSync(`git clone ${gitConfigurationUrl} ${configDirName}`, {stdio: [0, 1]});
+if(process.env.WIRE_CONFIGURATION_EXTERNAL_DIR === undefined) {
+  console.log(
+    `Loading configuration version "${gitConfigurationVersion}" for project "${pkg.name}" from "${gitConfigurationUrl}" \
+     and cleaning config directory "${configDir}"`);
+  fs.removeSync(configDir);
+  execSync(`git clone -b ${gitConfigurationVersion} ${gitConfigurationUrl} ${configDirName}`, {stdio: [0, 1]});
+} else {
+  console.log(`Using external config directory "${configDir}"`);
+}
 
 // Copy .env file configuration
 console.log('env', resolve(configDir, pkg.name, '.env'), resolve(root, '.env'));
